+++ src/Models/FormField.php
@@ -28,7 +28,8 @@ class FormField extends ActiveRecord
             ['name', 'required'],
             ['name', 'string'],
             ['form_id', 'integer'],
-            [['required', 'index_view'], 'integer', 'integerOnly' => true, 'min' => 0],
+            [['required', 'index_view', 'active'], 'integer', 'integerOnly' => true, 'min' => 0],
+            [['active'], 'default', 'value' => 1],
             [['required', 'index_view'], 'default', 'value' => 0],
             ['type', 'in', 'range' => self::FIELD_TYPES],
         ];